refactor: move MovieVault to the services package
